+++ Game.cpp
@@ -1,4 +1,5 @@
 #include "Game.h"
+#include "MemoryDebug.h"
 
 // For the DirectX Math library
 using namespace DirectX;
@@ -247,8 +248,8 @@ void Game::Update(float deltaTime, float totalTime)
 	//mouse pos
 	POINT cursorPos;
 	GetCursorPos(&cursorPos);
-	mouseX = cursorPos.x;
-	mouseY = cursorPos.y;
+	mouseX = static_cast<float>(cursorPos.x);
+	mouseY = static_cast<float>(cursorPos.y);
 
 
 	// Update camera
